+++ x/consensus/module.go
@@ -94,6 +94,7 @@ func init() {
 	appmodule.Register(
 		&modulev1.Module{},
 		appmodule.Provide(ProvideModule),
+		appmodule.Provide(ProvideAppVersionModifier),
 	)
 }
 
@@ -125,6 +126,7 @@ func ProvideModule(in ModuleInputs) ModuleOutputs {
 	m := NewAppModule(in.Cdc, k)
 	baseappOpt := func(app *baseapp.BaseApp) {
 		app.SetParamStore(k.ParamsStore)
+		app.SetVersionModifier(versionModifier{Keeper: k})
 	}
 
 	return ModuleOutputs{
@@ -132,4 +134,38 @@ func ProvideModule(in ModuleInputs) ModuleOutputs {
 		Module:        m,
 		BaseAppOption: baseappOpt,
 	}
+}
+
+type versionModifier struct {
+	Keeper keeper.Keeper
+}
+
+func (v versionModifier) SetAppVersion(ctx context.Context, version uint64) error {
+	params, err := v.Keeper.Params(ctx, nil)
+	if err != nil {
+		return err
+	}
+
+	updatedParams := params.Params
+	updatedParams.Version.App = version
+
+	err = v.Keeper.ParamsStore.Set(ctx, *updatedParams)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (v versionModifier) AppVersion(ctx context.Context) (uint64, error) {
+	params, err := v.Keeper.Params(ctx, nil)
+	if err != nil {
+		return 0, err
+	}
+
+	return params.Params.Version.GetApp(), nil
+}
+
+func ProvideAppVersionModifier(k keeper.Keeper) baseapp.VersionModifier {
+	return versionModifier{Keeper: k}
 }