+++ x/upgrade/keeper/migrations.go
@@ -2,6 +2,7 @@ package keeper
 
 import (
 	"encoding/binary"
+	"fmt"
 
 	storetypes "cosmossdk.io/core/store"
 	"cosmossdk.io/store/prefix"
@@ -9,6 +10,11 @@ import (
 
 	"github.com/cosmos/cosmos-sdk/runtime"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+)
+
+const (
+	// LegacyProtocolVersionByte was the prefix to look up Protocol Version (AppVersion)
+	LegacyProtocolVersionByte = 0x3
 )
 
 // Migrator is a struct for handling in-place store migrations.
@@ -46,4 +52,39 @@ func migrateDoneUpgradeKeys(ctx sdk.Context, storeService storetypes.KVStoreServ
 		oldDoneStore.Delete(oldKey)
 	}
 	return nil
+}
+
+// Migrate2to3 migrates from version 2 to 3.
+// It takes the legacy protocol version and if it exists, uses it to set
+// the app version (of the baseapp)
+func (m Migrator) Migrate2to3(ctx sdk.Context) error {
+	return migrateAppVersion(ctx, m.keeper)
+}
+
+func migrateAppVersion(ctx sdk.Context, keeper *Keeper) error {
+	if keeper.versionModifier == nil {
+		return fmt.Errorf("version modifier is not set")
+	}
+
+	store := keeper.storeService.OpenKVStore(ctx)
+	// if the key was never set then we don't need to migrate anything
+	exists, err := store.Has([]byte{LegacyProtocolVersionByte})
+	if err != nil {
+		return fmt.Errorf("error checking if legacy protocol version key exists: %w", err)
+	}
+	if !exists {
+		return nil
+	}
+
+	versionBytes, err := store.Get([]byte{LegacyProtocolVersionByte})
+	if err != nil {
+		return fmt.Errorf("error getting legacy protocol version: %w", err)
+	}
+	appVersion := binary.BigEndian.Uint64(versionBytes)
+
+	if err := keeper.versionModifier.SetAppVersion(ctx, appVersion); err != nil {
+		return fmt.Errorf("error migration app version: %w", err)
+	}
+
+	return store.Delete([]byte{LegacyProtocolVersionByte})
 }