+++ CHANGELOG.md
@@ -36,6 +36,18 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
+## [Unreleased Atom One Fork]
+
+### Features
+
+* (baseapp) [#1](https://github.com/atomone-hub/cosmos-sdk/pull/1) `SetProtocolVersion` has been renamed to `SetAppVersion`. It now updates the consensus params in baseapp's `ParamStore`.
+* (baseapp) [#1](https://github.com/atomone-hub/cosmos-sdk/pull/1) Re-add query router for custom abci queries.
+* (testing) [#1](https://github.com/atomone-hub/cosmos-sdk/pull/1) Add `MakeTestEncodingConfigWithOpts` to pass `testutil.CodecOptions` when creating `TestEncodingConfig`.
+* (x/auth) [#1](https://github.com/atomone-hub/cosmos-sdk/pull/1) Support legacy global AccountNumber.
+* (crypto/keyring) [#1](https://github.com/atomone-hub/cosmos-sdk/pull/1) Fix error wrapping in pkg crypto/keyring. Removed `cockroachdb/errors` in favour of `errorsmod.Wrap`.
+* (x/authz) [#1](https://github.com/atomone-hub/cosmos-sdk/pull/1) Added a limit of 200 grants pruned per `BeginBlock` and the `PruneExpiredGrants` message that prunes 75 expired grants on every run.
+* (x/consensus) [#1](https://github.com/atomone-hub/cosmos-sdk/pull/1) Address bug where sending `MsgUpdateParams` to `x/consensus` would overwrite `AppVersion` with zero value. Always use existing `AppVersion` param.
+
 ## [Unreleased]
 
 ### Improvements